Switch the default mode for clang-format to '-file'. Make 'LLVM' the fallback syntax used when we fail to find a '.clang-format' file. Adjust variable names appropriately.
Update the editor integration pieces that specify a '-style' option to
specify it as '-style=file'. I left the functionality in place because
even if the preferred method is to use '.clang-format' files, this way
if someone needs to clobber the style in their editor we show how to do
so in these examples.

Also check in a '.clang-format' file for Clang to ensure that separate
checkouts and builds of Clang from LLVM can still get the nice
formatting. =] This unfortunately required nuking the test for the
absence of a '.clang-format' file as now the directory happening to be
under your clang source tree will cause there to always be a file. ;]

git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk/tools/clang-format@189741 91177308-0d34-0410-b5e6-96231b3b80d8