clang-format-diff.py: Support -regex filter and more filename extensions
Add support for more filename extensions based on the list in the clang
plus JavaScript.

Also adds a -regex option so users can override defaults if they have unusual
file extensions or want to format everything in the diff.

Keeping with tradition the flag is modelled on Unix conventions, this time
matching the semantics of find(1).

git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk/tools/clang-format@196917 91177308-0d34-0410-b5e6-96231b3b80d8